Relocating unused unit tests to test-examples directory
