+++ package.json
@@ -50,7 +50,7 @@
     "docs": "./node_modules/apidoc/bin/apidoc -i src/ -o apidoc/",
     "lint": "node node_modules/eslint/bin/eslint.js src/ --fix",
     "clean": "node node_modules/depcheck/bin/depcheck.js",
-    "start:test": "mocha --timeout 1000000 --recursive --exit || true",
+    "start:test": "mocha --timeout 5000 --recursive --exit || true",
     "start:dev": "./node_modules/pm2/bin/pm2-dev start ./config/pm2.config.json --env dev && ./node_modules/pm2/bin/pm2 logs"
   }
 }
