+++ test-examples/test-cna-operations.js
@@ -4,7 +4,7 @@ chai.use(require('chai-uuid'))
 chai.use(chaiHttp)
 chai.should()
 
-var { TEST_CNA_UUID, TEST_USER_PUBLICKEY } = require('./test/constants')
+var { TEST_CNA_UUID, TEST_USER_PUBLICKEY } = require('../test/constants')
 
 describe('CNA Operations:', () => {
   it('Should Register User', function (done) {
