+++ test-examples/test-user-operations.js
@@ -8,7 +8,7 @@ chai.use(require('chai-string'))
 chai.use(chaiHttp)
 chai.should()
 
-var { TEST_USER_UUID, TEST_CVE_STATUS, TEST_CVE_ID } = require('./test/constants')
+var { TEST_USER_UUID, TEST_CVE_STATUS, TEST_CVE_ID } = require('../test/constants')
 
 describe('User Request CVE ID Operations:', () => {
   it('Should Request 2 CVE IDs', function (done) {
